Refactor example structure
